add nextBatch
