+++ velox/connectors/nexmark/Event.h
@@ -17,13 +17,418 @@
 
 #include <memory>
 
-#include "velox/connectors/nexmark/AuctionGenerator.h"
-#include "velox/connectors/nexmark/BidGenerator.h"
-#include "velox/connectors/nexmark/GeneratorConfig.h"
-#include "velox/connectors/nexmark/PersonGenerator.h"
+#include "velox/connectors/nexmark/NexmarkUtils.h"
+#include "velox/type/Type.h"
+#include "velox/type/StringView.h"
 #include "velox/vector/BaseVector.h"
+#include "velox/vector/FlatVector.h"
+#include "velox/vector/ComplexVector.h"
 
 namespace facebook::velox::connector::nexmark {
+
+/** A person either creating an auction or making a bid. */
+struct Person {
+ public:
+  /** Id of person. */
+  int64_t id; // primary key
+
+  /** Extra person properties. */
+  std::string name;
+  std::string emailAddress;
+  std::string creditCard;
+  std::string city;
+  std::string state;
+  int64_t dateTime; // unit: ms
+
+  /** Additional arbitrary payload for performance testing. */
+  std::string_view extra;
+
+  Person(
+      int64_t id,
+      std::string name,
+      std::string emailAddress,
+      std::string creditCard,
+      std::string city,
+      std::string state,
+      int64_t dateTime,
+      std::string_view extra)
+      : id(id),
+        name(std::move(name)),
+        emailAddress(std::move(emailAddress)),
+        creditCard(std::move(creditCard)),
+        city(std::move(city)),
+        state(std::move(state)),
+        dateTime(dateTime),
+        extra(std::move(extra)) {}
+
+  std::string toString() const {
+    return "Person{id=" + std::to_string(id) + ", name='" + name + '\'' +
+        ", emailAddress='" + emailAddress + '\'' + ", creditCard='" +
+        creditCard + '\'' + ", city='" + city + '\'' + ", state='" + state +
+        '\'' + ", dateTime=" + formatDateTime(dateTime) + ", extra='" +
+        std::string(extra) + '\'' + '}';
+  }
+
+  // Person RowType
+  static TypePtr createType() {
+    return ROW(
+        {
+            "id", // Person ID
+            "name", // Name
+            "emailAddress", // Email
+            "creditCard", // Credit Card
+            "city", // City
+            "state", // State/Province
+            "dateTime", // State/Province
+            "extra" // Extra Information
+        },
+        {
+            BIGINT(), // id
+            VARCHAR(), // name
+            VARCHAR(), // email
+            VARCHAR(), // creditCard
+            VARCHAR(), // city
+            VARCHAR(), // state
+            TIMESTAMP(), // dateTime
+            VARCHAR() // extra
+        });
+  }
+
+  static RowVectorPtr createVector(int rows, memory::MemoryPool* pool) {
+    auto idVector = BaseVector::create(BIGINT(), rows, pool);
+    auto nameVector = BaseVector::create(VARCHAR(), rows, pool);
+    auto emailVector = BaseVector::create(VARCHAR(), rows, pool);
+    auto creditCardVector = BaseVector::create(VARCHAR(), rows, pool);
+    auto cityVector = BaseVector::create(VARCHAR(), rows, pool);
+    auto stateVector = BaseVector::create(VARCHAR(), rows, pool);
+    auto dateTimeVector = BaseVector::create(TIMESTAMP(), rows, pool);
+    auto extraVector = BaseVector::create(VARCHAR(), rows, pool);
+
+    return std::make_shared<RowVector>(
+        pool,
+        createType(),
+        nullptr,
+        rows,
+        std::vector<VectorPtr>{
+            idVector,
+            nameVector,
+            emailVector,
+            creditCardVector,
+            cityVector,
+            stateVector,
+            dateTimeVector,
+            extraVector});
+  }
+
+  static void
+  fillVector(RowVector* personVector, int index, const Person* person) {
+    if (!person) {
+      personVector->setNull(index, true);
+      return;
+    }
+
+    auto idVector = personVector->childAt(0)->asFlatVector<int64_t>();
+    auto nameVector = personVector->childAt(1)->asFlatVector<StringView>();
+    auto emailVector = personVector->childAt(2)->asFlatVector<StringView>();
+    auto creditCardVector =
+        personVector->childAt(3)->asFlatVector<StringView>();
+    auto cityVector = personVector->childAt(4)->asFlatVector<StringView>();
+    auto stateVector = personVector->childAt(5)->asFlatVector<StringView>();
+    auto dateTimeVector = personVector->childAt(6)->asFlatVector<Timestamp>();
+    auto extraVector = personVector->childAt(7)->asFlatVector<StringView>();
+
+    idVector->set(index, person->id);
+    nameVector->set(index, StringView(person->name));
+    emailVector->set(index, StringView(person->emailAddress));
+    creditCardVector->set(index, StringView(person->creditCard));
+    cityVector->set(index, StringView(person->city));
+    stateVector->set(index, StringView(person->state));
+    dateTimeVector->set(
+        index,
+        Timestamp(person->dateTime / 1000, (person->dateTime) % 1000 * 1000));
+    extraVector->set(index, StringView(person->extra));
+  }
+};
+
+/** An auction submitted by a person. */
+struct Auction {
+  /** Id of auction. */
+  int64_t id; // primary key
+
+  /** Extra auction properties. */
+  std::string itemName;
+
+  std::string description;
+
+  /** Initial bid price, in cents. */
+  int64_t initialBid;
+
+  /** Reserve price, in cents. */
+  int64_t reserve;
+
+  int64_t dateTime;
+
+  /** When does auction expire? (ms since epoch). Bids at or after this time are
+   * ignored. */
+  int64_t expires;
+
+  /** Id of person who instigated auction. */
+  int64_t seller; // foreign key: Person.id
+
+  /** Id of category auction is listed under. */
+  int64_t category; // foreign key: Category.id
+
+  /** Additional arbitrary payload for performance testing. */
+  std::string_view extra;
+
+  /** Constructor with all fields */
+  Auction(
+      int64_t id,
+      std::string itemName,
+      std::string description,
+      int64_t initialBid,
+      int64_t reserve,
+      int64_t dateTime,
+      int64_t expires,
+      int64_t seller,
+      int64_t category,
+      std::string_view extra)
+      : id(id),
+        itemName(std::move(itemName)),
+        description(std::move(description)),
+        initialBid(initialBid),
+        reserve(reserve),
+        dateTime(dateTime),
+        expires(expires),
+        seller(seller),
+        category(category),
+        extra(std::move(extra)) {}
+
+  std::string toString() const {
+    return "Auction{id=" + std::to_string(id) + ", itemName='" + itemName +
+        '\'' + ", description='" + description + '\'' +
+        ", initialBid=" + std::to_string(initialBid) +
+        ", reserve=" + std::to_string(reserve) +
+        ", dateTime=" + formatDateTime(dateTime) +
+        ", expires=" + formatDateTime(expires) +
+        ", seller=" + std::to_string(seller) +
+        ", category=" + std::to_string(category) + ", extra='" +
+        std::string(extra) + '\'' + '}';
+  }
+
+  // Auction RowType
+  static TypePtr createType() {
+    return ROW(
+        {
+            "id", // Auction ID
+            "name", // Auction name
+            "description", // Description
+            "initialBid", // Initial bid
+            "reserve", // Reserve price
+            "dateTime", // Timestamp
+            "expires", // Expiration time
+            "seller", // Seller ID
+            "category", // Category
+            "extra" // Additional information
+        },
+        {
+            BIGINT(), // id
+            VARCHAR(), // name
+            VARCHAR(), // description
+            BIGINT(), // initialBid
+            BIGINT(), // reserve
+            TIMESTAMP(), // timestamp
+            TIMESTAMP(), // expires
+            BIGINT(), // seller
+            BIGINT(), // category
+            VARCHAR() // extra
+        });
+  }
+
+  static RowVectorPtr createVector(int rows, memory::MemoryPool* pool) {
+    auto idVector = BaseVector::create(BIGINT(), rows, pool);
+    auto nameVector = BaseVector::create(VARCHAR(), rows, pool);
+    auto descVector = BaseVector::create(VARCHAR(), rows, pool);
+    auto initialBidVector = BaseVector::create(BIGINT(), rows, pool);
+    auto reserveVector = BaseVector::create(BIGINT(), rows, pool);
+    auto dateTimeVector = BaseVector::create(TIMESTAMP(), rows, pool);
+    auto expiresVector = BaseVector::create(TIMESTAMP(), rows, pool);
+    auto sellerVector = BaseVector::create(BIGINT(), rows, pool);
+    auto categoryVector = BaseVector::create(BIGINT(), rows, pool);
+    auto extraVector = BaseVector::create(VARCHAR(), rows, pool);
+
+    return std::make_shared<RowVector>(
+        pool,
+        createType(),
+        nullptr,
+        rows,
+        std::vector<VectorPtr>{
+            idVector,
+            nameVector,
+            descVector,
+            initialBidVector,
+            reserveVector,
+            dateTimeVector,
+            expiresVector,
+            sellerVector,
+            categoryVector,
+            extraVector});
+  }
+
+  static void
+  fillVector(RowVector* auctionVector, int index, const Auction* auction) {
+    if (!auction) {
+      auctionVector->setNull(index, true);
+      return;
+    }
+
+    auto idVector = auctionVector->childAt(0)->asFlatVector<int64_t>();
+    auto nameVector = auctionVector->childAt(1)->asFlatVector<StringView>();
+    auto descVector = auctionVector->childAt(2)->asFlatVector<StringView>();
+    auto initialBidVector = auctionVector->childAt(3)->asFlatVector<int64_t>();
+    auto reserveVector = auctionVector->childAt(4)->asFlatVector<int64_t>();
+    auto dateTimeVector = auctionVector->childAt(5)->asFlatVector<Timestamp>();
+    auto expiresVector = auctionVector->childAt(6)->asFlatVector<Timestamp>();
+    auto sellerVector = auctionVector->childAt(7)->asFlatVector<int64_t>();
+    auto categoryVector = auctionVector->childAt(8)->asFlatVector<int64_t>();
+    auto extraVector = auctionVector->childAt(9)->asFlatVector<StringView>();
+
+    idVector->set(index, auction->id);
+    nameVector->set(index, StringView(auction->itemName));
+    descVector->set(index, StringView(auction->description));
+    initialBidVector->set(index, auction->initialBid);
+    reserveVector->set(index, auction->reserve);
+    dateTimeVector->set(
+        index,
+        Timestamp(auction->dateTime / 1000, (auction->dateTime) % 1000 * 1000));
+    expiresVector->set(
+        index,
+        Timestamp(auction->expires / 1000, (auction->expires) % 1000 * 1000));
+    sellerVector->set(index, auction->seller);
+    categoryVector->set(index, auction->category);
+    extraVector->set(index, StringView(auction->extra));
+  }
+};
+
+struct Bid {
+  /** Id of auction this bid is for. */
+  int64_t auction; // foreign key: Auction.id
+
+  /** Id of person bidding in auction. */
+  int64_t bidder; // foreign key: Person.id
+
+  /** Price of bid, in cents. */
+  int64_t price;
+
+  /** The channel introduced this bidding. */
+  std::string channel;
+
+  /** The url of this bid. */
+  std::string url;
+
+  /**
+   * Instant at which bid was made (ms since epoch). NOTE: This may be earlier
+   * than the system's event time.
+   */
+  int64_t dateTime;
+
+  /** Additional arbitrary payload for performance testing. */
+  std::string_view extra;
+
+  Bid(int64_t auction,
+      int64_t bidder,
+      int64_t price,
+      std::string channel,
+      std::string url,
+      int64_t dateTime,
+      std::string_view extra)
+      : auction(auction),
+        bidder(bidder),
+        price(price),
+        channel(std::move(channel)),
+        url(std::move(url)),
+        dateTime(dateTime),
+        extra(std::move(extra)) {}
+
+  std::string toString() const {
+    return "Bid{auction=" + std::to_string(auction) +
+        ", bidder=" + std::to_string(bidder) +
+        ", price=" + std::to_string(price) + ", channel=" + channel +
+        ", url=" + url + ", dateTime=" + formatDateTime(dateTime) +
+        ", extra='" + std::string(extra) + '\'' + '}';
+  }
+
+  // Bid RowType
+  static TypePtr createType() {
+    return ROW(
+        {
+            "auction", // Auction ID
+            "bidder", // Bidder ID
+            "price", // Price
+            "channel", // Channel
+            "url", // Url
+            "dateTime", // Timestamp
+            "extra" // Additional information
+        },
+        {
+            BIGINT(), // auction
+            BIGINT(), // bidder
+            BIGINT(), // price
+            VARCHAR(), // channel
+            VARCHAR(), // url
+            TIMESTAMP(), // timestamp
+            VARCHAR() // extra
+        });
+  }
+
+  static RowVectorPtr createVector(int rows, memory::MemoryPool* pool) {
+    auto auctionVector = BaseVector::create(BIGINT(), rows, pool);
+    auto bidderVector = BaseVector::create(BIGINT(), rows, pool);
+    auto priceVector = BaseVector::create(BIGINT(), rows, pool);
+    auto channelVector = BaseVector::create(VARCHAR(), rows, pool);
+    auto urlVector = BaseVector::create(VARCHAR(), rows, pool);
+    auto dateTimeVector = BaseVector::create(TIMESTAMP(), rows, pool);
+    auto extraVector = BaseVector::create(VARCHAR(), rows, pool);
+
+    return std::make_shared<RowVector>(
+        pool,
+        createType(),
+        nullptr,
+        rows,
+        std::vector<VectorPtr>{
+            auctionVector,
+            bidderVector,
+            priceVector,
+            channelVector,
+            urlVector,
+            dateTimeVector,
+            extraVector});
+  }
+
+  static void fillVector(RowVector* bidVector, int index, const Bid* bid) {
+    if (!bid) {
+      bidVector->setNull(index, true);
+      return;
+    }
+
+    auto auctionVector = bidVector->childAt(0)->asFlatVector<int64_t>();
+    auto bidderVector = bidVector->childAt(1)->asFlatVector<int64_t>();
+    auto priceVector = bidVector->childAt(2)->asFlatVector<int64_t>();
+    auto channelVector = bidVector->childAt(3)->asFlatVector<StringView>();
+    auto urlVector = bidVector->childAt(4)->asFlatVector<StringView>();
+    auto dateTimeVector = bidVector->childAt(5)->asFlatVector<Timestamp>();
+    auto extraVector = bidVector->childAt(6)->asFlatVector<StringView>();
+
+    auctionVector->set(index, bid->auction);
+    bidderVector->set(index, bid->bidder);
+    priceVector->set(index, bid->price);
+    channelVector->set(index, StringView(bid->channel));
+    urlVector->set(index, StringView(bid->url));
+    dateTimeVector->set(
+        index, Timestamp(bid->dateTime / 1000, (bid->dateTime) % 1000 * 1000));
+    extraVector->set(index, StringView(bid->extra));
+  }
+};
 
 /// Represents an event containing either a Person, an Auction, or a Bid.
 class Event {
